backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
aioboto3==13.3.0
python-dotenv==1.0.1
# 認証関連
google-auth==2.38.0
python-jose[cryptography]==3.4.0
httpx==0.28.1
